Main menu changes...
Imported a new resource for the navigation and simpler react structure: "react-bootstrap"
Refactoring the menu with the new resource.
Rearange the menu layout with CSS.
Replace the Profile component to SigningStatus.
Setup the basic logic of the signingStatus apperance.
Create a SigningModal component for signing in/up.
Create a nicer look to the SearchField component and create function to the  button.